♻️ refactor: 백신 조회 고정된 컴포넌트들 temp로 코드 분리
